Fix build error introduced in 5f701793f63f16d111db3e3d2e91134fcc179c5b.
* linux/src/include/linux/interrupt.h (intr_count): Declare as unsigned int.

+++ linux/src/include/linux/interrupt.h
@@ -14,7 +14,7 @@ struct irqaction {
 	struct irqaction *next;
 };
 
-extern unsigned long intr_count;
+extern unsigned int intr_count;
 
 extern int bh_mask_count[32];
 extern unsigned int bh_active;
